+++ src/Time.cpp
@@ -1,34 +1,32 @@
 //
 // Created by OSAMA ASHRAF on 4/6/2024.
 //
-#include "../headers/Time.h"
+#include "headers/Time.h"
 #include <sstream>
 #include <vector>
 #include <algorithm>
 
 using namespace std;
 // Global Functions
-void toggle(string &period)
-{
-    if (period == "AM")
-        period = "PM";
-    else
-        period = "AM";
+void toggle(string &period) {
+    if (period == "AM") period = "PM";
+    else period = "AM";
 }
-string timeFormat(int x, bool isHour = false)
-{
+string timeFormat(int x, bool isHour = false){
     ostringstream oss;
-    if (x < 10 && !isHour)
+    if(x < 10 && !isHour)
         oss << 0;
-    oss << x;
+    oss<< x;
 
     return oss.str();
 }
 
 // Constructors
-Time::Time(int hours, int minutes, int seconds, string period) : seconds(seconds), minutes(minutes), hours(hours), period(period) {}
+Time::Time(int hours, int minutes, int seconds, string period) :
+        seconds(seconds), minutes(minutes), hours(hours), period(period) {}
 
 Time::Time() : Time(0, 0, 0, "AM") {}
+
 
 // Getters
 int Time::getSeconds() const { return seconds; }
@@ -39,6 +37,7 @@ int Time::getHours() const { return hours; }
 
 string Time::getPeriod() const { return period; }
 
+
 // Setters
 bool Time::setSeconds(int s) { return !(s > 59 || s < 1) && (seconds = s); }
 
@@ -48,33 +47,29 @@ bool Time::setHours(int h) { return !(h > 24 || h < 1) && (hours = h); }
 
 bool Time::setPeriod(string p) { return (p == "AM" || p == "PM") && (period = p, true); }
 
+
 // Methods
-bool Time::isValid()
-{
+bool Time::isValid() {
     if (setSeconds(seconds) && setMinutes(minutes) && setHours(hours) && setPeriod(period))
         return true;
 
     return false;
 }
 
-void Time::addSeconds(int s)
-{
+void Time::addSeconds(int s) {
     int sum = seconds + s;
 
     seconds = sum % 60;
     addMinutes(sum / 60);
 }
 
-void Time::addMinutes(int m)
-{
+void Time::addMinutes(int m) {
     addHours((minutes + m) / 60);
     minutes = (minutes + m) % 60;
 }
 
-void Time::addHours(int h)
-{
-    while (h--)
-    {
+void Time::addHours(int h) {
+    while (h--) {
         hours++;
 
         if (hours == 12)
@@ -85,8 +80,7 @@ void Time::addHours(int h)
     }
 }
 
-string Time::toString(string format, string separator)
-{
+string Time::toString(string format, string separator) {
     vector<pair<int, string>> index;
 
     int h_index = format.find("hh"), m_index = format.find("mm"), s_index = format.find("ss");
@@ -101,8 +95,8 @@ string Time::toString(string format, string separator)
     sort(index.begin(), index.end());
 
     ostringstream oss;
-    for (int i = 0; i < index.size(); i++)
-        oss << index[i].second << (i < index.size() - 1 ? separator : " ");
+    for (int i = 0; i < (int)index.size(); i++)
+        oss << index[i].second << (i < (int)index.size() - 1 ? separator : " ");
     oss << period;
 
     return oss.str();
